+++ src/screens/Statistics/HighlightsSection.tsx
@@ -19,6 +19,37 @@ import { Subtitle } from "./Subtitle";
 import { TagPeaksCard } from "./TagPeaksCards";
 import { TagsDistributionCard } from "./TagsDistributionCard";
 import { Title } from "./Title";
+import { EmotionsDistributionCard } from "./EmotionsDistributionCard";
+
+const EmptryState = () => {
+  const colors = useColors();
+
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        marginTop: 16,
+        padding: 16,
+        paddingVertical: 32,
+        borderWidth: 1,
+        borderColor: colors.statisticsNoDataBorder,
+        borderStyle: "dashed",
+      }}
+    >
+      <Text
+        style={{
+          color: colors.textSecondary,
+          fontSize: 17,
+          textAlign: "center",
+        }}
+      >
+        {t("statistics_no_highlights")}
+      </Text>
+    </View>
+  )
+}
 
 export const HighlightsSection = ({ items }: { items: LogItem[] }) => {
   const colors = useColors();
@@ -27,24 +58,12 @@ export const HighlightsSection = ({ items }: { items: LogItem[] }) => {
   const statistics = useStatistics();
   const logState = useLogState();
 
-  const showMoodAvg =
-    statistics.isAvailable("mood_avg") &&
-    statistics.state.moodAvgData.ratingHighestPercentage > 60;
-
-  const showMoodPeaksPositve =
-    statistics.isAvailable("mood_peaks_positive") &&
-    statistics.state.moodPeaksPositiveData.days.length >= 2;
-
-  const showMoodPeaksNegative =
-    statistics.isAvailable("mood_peaks_negative") &&
-    statistics.state.moodPeaksNegativeData.days.length >= 2;
-
-  const showTagPeaks =
-    statistics.isAvailable("tags_peaks") &&
-    statistics.state.tagsPeaksData.tags.filter((tag) => tag.items.length > 5).length > 0;
-
+  const showMoodAvg = statistics.isHighlighted("mood_avg");
+  const showMoodPeaksPositve = statistics.isHighlighted("mood_peaks_positive");
+  const showMoodPeaksNegative = statistics.isHighlighted("mood_peaks_negative");
+  const showTagPeaks = statistics.isHighlighted("tags_peaks");
   const showTagsDistribution = statistics.isAvailable("tags_distribution")
-
+  const showEmotionsDistribution = statistics.isAvailable("emotions_distribution")
   const showMoodChart = logState.items.filter((item) => dayjs(item.dateTime).isAfter(dayjs().subtract(14, "day"))).length >= 4
 
   useEffect(() => {
@@ -65,13 +84,16 @@ export const HighlightsSection = ({ items }: { items: LogItem[] }) => {
       tags_distribution_item_count?: number
       mood_chart_show: boolean;
       mood_chart_item_count?: number
+      emotions_distribution_show: boolean;
+      emotions_distribution_item_count?: number
     } = {
       mood_avg_show: showMoodAvg,
       mood_peaks_positive_show: showMoodPeaksPositve,
       mood_peaks_negative_show: showMoodPeaksNegative,
       tags_peaks_show: showTagPeaks,
       tags_distribution_show: showTagsDistribution,
-      mood_chart_show: showMoodChart
+      mood_chart_show: showMoodChart,
+      emotions_distribution_show: showEmotionsDistribution,
     }
 
     if (showMoodAvg) {
@@ -92,6 +114,9 @@ export const HighlightsSection = ({ items }: { items: LogItem[] }) => {
     }
     if (showMoodChart) {
       cards.mood_chart_item_count = logState.items.filter((item) => dayjs(item.dateTime).isAfter(dayjs().subtract(14, "day"))).length
+    }
+    if (showEmotionsDistribution) {
+      cards.emotions_distribution_item_count = statistics.state.emotionsDistributionData.emotions.length
     }
 
     analytics.track('statistics_relevant_highlights', {
@@ -118,35 +143,19 @@ export const HighlightsSection = ({ items }: { items: LogItem[] }) => {
           !showTagsDistribution &&
           !showMoodChart
         ) && (
-            <View
-              style={{
-                flex: 1,
-                justifyContent: "center",
-                alignItems: "center",
-                marginTop: 16,
-                padding: 16,
-                paddingVertical: 32,
-                borderWidth: 1,
-                borderColor: colors.statisticsNoDataBorder,
-                borderStyle: "dashed",
-              }}
-            >
-              <Text
-                style={{
-                  color: colors.textSecondary,
-                  fontSize: 17,
-                  textAlign: "center",
-                }}
-              >
-                {t("statistics_no_highlights")}
-              </Text>
-            </View>
+            <EmptryState />
           )}
 
         {showMoodChart && (
           <MoodChart
             title={t("statistics_mood_chart_highlights_title")}
             startDate={dayjs().subtract(14, "days").format(DATE_FORMAT)}
+          />
+        )}
+
+        {showMoodChart && (
+          <EmotionsDistributionCard
+            data={statistics.state.emotionsDistributionData}
           />
         )}